feat: filter form item settings
