+++ packages/core/client/src/schema-component/antd/form-item/index.tsx
@@ -1,2 +1,3 @@
 export * from './FormItem';
 export * from './FormItem.Settings';
+export * from './FormItem.FilterFormSettings';
